fix review

+++ tasks/seq/chizhov_m_trapezoid_method/include/ops_seq.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <stddef.h>
 #include <functional>
 #include <utility>
 #include <vector>
@@ -9,7 +10,7 @@
 namespace chizhov_m_trapezoid_method_seq {
 using Function = std::function<double(const std::vector<double>&)>;
 
-double TrapezoidMethod(Function& f, int div, int dim, std::vector<double>& lower_limits,
+double TrapezoidMethod(Function& f, size_t div, size_t dim, std::vector<double>& lower_limits,
                        std::vector<double>& upper_limits);
 
 class TestTaskSequential : public ppc::core::Task {
@@ -25,8 +26,8 @@ class TestTaskSequential : public ppc::core::Task {
   Function f_;
   std::vector<double> lower_limits_;
   std::vector<double> upper_limits_;
-  int div_;
-  int dim_;
+  size_t div_;
+  size_t dim_;
   double res_;
 };
 }  // namespace chizhov_m_trapezoid_method_seq
